+++ sdgym/__init__.py
@@ -8,12 +8,18 @@ __author__ = 'MIT Data To AI Lab'
 __copyright__ = 'Copyright (c) 2018, MIT Data To AI Lab'
 __email__ = 'dailabmit@gmail.com'
 __license__ = 'MIT'
-__version__ = '0.2.1'
+__version__ = '0.2.2.dev1'
 
-from sdgym.benchmark import benchmark
+from sdgym import benchmark, results, synthesizers
+from sdgym.benchmark import run
 from sdgym.data import load_dataset
+from sdgym.synthesizers import get_all_synthesizers
 
 __all__ = [
     'benchmark',
-    'load_dataset'
+    'synthesizers',
+    'results',
+    'run',
+    'load_dataset',
+    'get_all_synthesizers',
 ]