+++ sdgym/data.py
@@ -1,10 +1,13 @@
 import json
+import logging
 import os
 import urllib
 
 import numpy as np
 
 from sdgym.constants import CATEGORICAL, ORDINAL
+
+LOGGER = logging.getLogger(__name__)
 
 BASE_URL = 'http://sdgym.s3.amazonaws.com/datasets/'
 DATA_PATH = os.path.join(os.path.dirname(__file__), 'data')
@@ -19,7 +22,10 @@ def _load_file(filename, loader):
     local_path = os.path.join(DATA_PATH, filename)
     if not os.path.exists(local_path):
         os.makedirs(DATA_PATH, exist_ok=True)
-        urllib.request.urlretrieve(BASE_URL + filename, local_path)
+        url = BASE_URL + filename
+
+        LOGGER.info('Downloading file %s to %s', url, local_path)
+        urllib.request.urlretrieve(url, local_path)
 
     return loader(local_path)
 
@@ -37,6 +43,7 @@ def _get_columns(metadata):
 
 
 def load_dataset(name, benchmark=False):
+    LOGGER.info('Loading dataset %s', name)
     data = _load_file(name + '.npz', np.load)
     meta = _load_file(name + '.json', _load_json)